made app.js clearner
